Updates unittests and mocks for filter-functions and db-functions
